Refactored and Fixed tests
